+++ src/collection.test.ts
@@ -1,24 +1,46 @@
 import { ok, error } from "./result";
-import { allOk, collectOks, firstOk } from "./collection";
+import { allOk, allOkAsync, firstOk } from "./collection";
 
 describe("collection", () => {
   describe("allOk", () => {
-    it("returns an :ok if all results are :ok", () => {
-      expect(allOk([ok(1), ok(2), ok(3)])).toEqual(ok(null));
+    it("returns all :ok messages if all results are :ok", () => {
+      expect(allOk([ok(1), ok(2), ok(3)])).toEqual(ok([1, 2, 3]));
     });
 
     it("returns the first :error if any results are :error", () => {
-      expect(allOk([ok(1), error(2), ok(3)])).toEqual(error(2));
+      expect(allOk([ok(1), error(2), error(3), ok(4)])).toEqual(error(2));
     });
   });
 
-  describe("collectOks", () => {
-    it("returns all :ok messages if all results are :ok", () => {
-      expect(collectOks([ok(1), ok(2), ok(3)])).toEqual(ok([1, 2, 3]));
+  describe("allOkAsync", () => {
+    it("returns all ok messages if all results are ok", async () => {
+      expect(
+        await allOkAsync([
+          Promise.resolve(ok(1)),
+          Promise.resolve(ok(2)),
+          Promise.resolve(ok(3))
+        ])
+      ).toEqual(ok([1, 2, 3]));
     });
 
-    it("returns the first :error if any results are :error", () => {
-      expect(collectOks([ok(1), error(2), ok(3)])).toEqual(error(2));
+    it("returns the first error if any results are error", async () => {
+      expect(
+        await allOkAsync([
+          Promise.resolve(ok(1)),
+          Promise.resolve(error(2)),
+          Promise.resolve(ok(3))
+        ])
+      ).toEqual(error(2));
+    });
+
+    it("returns the first error if any promises are rejected", async () => {
+      expect(
+        await allOkAsync([
+          Promise.resolve(ok(1)),
+          Promise.reject(ok(2)),
+          Promise.resolve(ok(3))
+        ])
+      ).toEqual(error(ok(2)));
     });
   });