+++ src/pipeAsync.test.ts
@@ -1,7 +1,7 @@
 import { pipeAsync, createPipeAsync } from "./pipeAsync";
 
 describe("asyncPipe", () => {
-  let asyncAddOne: jest.Mock<number, [number]>;
+  let asyncAddOne: jest.Mock<Promise<number>, [number]>;
 
   describe("pipeAsync", () => {
     beforeEach(() => {
@@ -11,12 +11,26 @@ describe("asyncPipe", () => {
     });
 
     it("does things async", async () => {
-      await pipeAsync(1, asyncAddOne, asyncAddOne);
+      const answer = await pipeAsync(1, asyncAddOne, asyncAddOne);
+
+      expect(answer).toBe(3);
 
       expect(asyncAddOne).toHaveBeenCalledTimes(2);
 
       expect(asyncAddOne).toHaveBeenNthCalledWith(1, 1);
       expect(asyncAddOne).toHaveBeenNthCalledWith(2, 2);
+    });
+
+    it("works with a starting promise", async () => {
+      const answer = await pipeAsync(
+        Promise.resolve(1),
+        asyncAddOne,
+        asyncAddOne
+      );
+
+      expect(answer).toBe(3);
+
+      expect(asyncAddOne).toHaveBeenCalledTimes(2);
     });
   });