+++ src/sync.ts
@@ -1,17 +1,29 @@
 import { Result, ok, error, isOk, isError } from "./result";
 
 /**
- * Edits a value that's wrapped in an {ok: data}
+ * Transforms a successful result, and passes through a failed result.
  *
- * Takes a Result and a mapping function.
+ * Takes a mapping function, then a result.
  * If the result is an Ok, applies the function to the data.
  * If the result is an Error, passes the Result through unchanged.
  *
- * Wraps the return value of f in an {ok: new_data}.
+ * Wraps the output of the function in an Ok.
+ *
  * @param f - the function to run on the ok data
  * @param result - The result to match against
+ *
+ * ```javascript
+ * function add1(n) {
+ *   return n + 1
+ * }
+ *
+ * okThen(add1)(ok(1)) // ok(2)
+ * okThen(add1)(error("bad")) // error("bad")
+ * okThen(add1)(ok(-1)) // ok(0)
+ * ```
  */
-export function ifOk<OkData, ErrorMessage, OkOutput>(
+
+export function okThen<OkData, ErrorMessage, OkOutput>(
   f: (ok: OkData) => OkOutput
 ) {
   return (
@@ -26,82 +38,28 @@ export function ifOk<OkData, ErrorMessage, OkOutput>(
 }
 
 /**
- * Runs a function for side effects on the payload, only if the result is Ok.
+ * Transforms a failed result, and passes through a successful result.
  *
- * @param f - the function to run on the ok data
- * @param result - The result to match against
- * @returns The original result
- */
-export function okSideEffect<OkData, ErrorMessage>(
-  f: (ok: OkData) => any
-): (result: Result<OkData, ErrorMessage>) => Result<OkData, ErrorMessage> {
-  return ifOk((data: OkData) => {
-    f(data);
-    return data;
-  });
-}
-
-/**
- * Chains together operations that may succeed or fail
- *
- * Takes a Result and a mapping function.
- * If the result is an Ok, applies the function to the data.
- * If the result is an Error, passes the Result through unchanged.
- *
- * @param f - the function to run on the ok data
- * @param result - The result to match against
- * @returns - The Result from function f or the Error
- */
-export function chainOk<OkData, ErrorMessage, OkOutput, ErrorOutput>(
-  f: (ok: OkData) => Result<OkOutput, ErrorOutput>
-) {
-  return (
-    result: Result<OkData, ErrorMessage>
-  ): Result<OkOutput, ErrorMessage | ErrorOutput> => {
-    if (isError(result)) {
-      return result;
-    }
-
-    return f(result.ok);
-  };
-}
-
-/**
- * Replaces a value that's wrapped in an {ok: data}
- * Useful if you don't care about the data, just the fact that previous call succeeded.
- *
- * Takes a Result and a mapping function.
- * If the result is an Error, applies the function to the message.
+ * Takes a mapping function, then a result.
+ * If the result is an Error, applies the function to the data.
  * If the result is an Ok, passes the Result through unchanged.
  *
- * It wraps the return value in an {error: new_message}.
- * @param f - the function to run on the ok data
- * @param result - The result to match against
- */
-export function replaceOk<OkData, ErrorMessage, OkOutput>(newData: OkOutput) {
-  return (
-    result: Result<OkData, ErrorMessage>
-  ): Result<OkOutput, ErrorMessage> => {
-    if (isError(result)) {
-      return result;
-    }
-
-    return ok(newData);
-  };
-}
-
-/**
- * Edits a value that's wrapped in an {error: data}
+ * Wraps the output of the function in an Error.
  *
- * Takes a Result and a mapping function.
- * If the result is an Error, applies the function to the message.
- * If the result is an Ok, passes the Result through unchanged.
- *
- * It wraps the return value in an {error: new_message}.
- * @param f - the function to run on the ok data
+ * @param f - the function to run on the error data
  * @param result - The result to match against
+ * @returns - The Result from function f or the Ok result
+ *
+ * ```javascript
+ * const normalizeErrorCase = errorThen(message =>
+ *   message.toLowerCase()
+ * )
+ *
+ * normalizeErrorCase(ok("alice")) // ok("alice")
+ * normalizeErrorCase(error("NOT FOUND")) // error("not found")
+ * ```
  */
-export function ifError<OkData, ErrorMessage, ErrorOutput>(
+export function errorThen<OkData, ErrorMessage, ErrorOutput>(
   f: (error: ErrorMessage) => ErrorOutput
 ) {
   return (
@@ -116,33 +74,67 @@ export function ifError<OkData, ErrorMessage, ErrorOutput>(
 }
 
 /**
- * Runs a function for side effects on the payload, only if the result is Error.
+ * Performs an operation that could succeed or fail on a successful result
+ * Passes through a failed result.
  *
- * @param f - the function to run on the error message
+ * Takes a mapping function, then a result.
+ * If the result is an Ok, applies the function to the data.
+ * If the result is an Error, passes the Result through unchanged.
+ *
+ * The return value must itself be a result, which will be returned.
+ * @returns - The Result from function f or the Error
+ *
+ * @param f - the function to run on the ok data
  * @param result - The result to match against
- * @returns The original result
+ *
+ * ```javascript
+ * function positiveAdder(n) {
+ *   return n < 0 ? error("only positive") : ok(n + 1);
+ * }
+ *
+ * okThen(positiveAdder)(ok(1)) // ok(2)
+ * okThen(positiveAdder)(error("bad")) // error("bad")
+ * okThen(positiveAdder)(ok(-1)) // error("only positive")
+ * ```
  */
-export function errorSideEffect<OkData, ErrorMessage>(
-  f: (error: ErrorMessage) => any
-): (result: Result<OkData, ErrorMessage>) => Result<OkData, ErrorMessage> {
-  return ifError((message: ErrorMessage) => {
-    f(message);
-    return message;
-  });
+export function okChain<OkData, ErrorMessage, OkOutput, ErrorOutput>(
+  f: (ok: OkData) => Result<OkOutput, ErrorOutput>
+) {
+  return (
+    result: Result<OkData, ErrorMessage>
+  ): Result<OkOutput, ErrorMessage | ErrorOutput> => {
+    if (isError(result)) {
+      return result;
+    }
+
+    return f(result.ok);
+  };
 }
 
 /**
- * Chains together operations that may succeed or fail
+ * Attempts to rescue failed results. Passes successful ones through.
  *
- * Takes a Result and a mapping function.
- * If the result is an Error, applies the function to the data and returns the new result.
+ * Takes a mapping function, then a result.
+ * If the result is an Error, applies the function to the data.
  * If the result is an Ok, passes the Result through unchanged.
  *
- * @param f - the function to run on the error message
+ * The return value must itself be a result, which will be returned.
+ *
+ * @param f - the function to run on the error data
  * @param result - The result to match against
  * @returns - The Result from function f or the Ok result
+ *
+ * ```javascript
+ * const rescueNotFound = errorThen(errorMessage =>
+ *   errorMessage === "not found" ? ok("unknown") : error(errorMessage)
+ * );
+ *
+ * rescueNotFound(ok("alice")) // ok("alice")
+ * rescueNotFound(error("not found")) // ok("unknown")
+ * rescueNotFound(error("network error")) // error("network error")
+ * ```
  */
-export function chainError<OkData, ErrorMessage, OkOutput, ErrorOutput>(
+export function errorRescue<OkData, ErrorMessage, OkOutput, ErrorOutput>(
   f: (ok: ErrorMessage) => Result<OkOutput, ErrorOutput>
 ) {
   return (
@@ -157,8 +149,56 @@ export function chainError<OkData, ErrorMessage, OkOutput, ErrorOutput>(
 }
 
 /**
- * Replaces a value that's wrapped in an {error: data}
- * Useful if you don't care about the data, just the fact that previous call failed.
+ * Runs a function for side effects on the payload, only if the result is Ok.
+ *
+ * @param f - the function to run on the ok data
+ * @param result - The result to match against
+ * @returns The original result
+ *
+ * ```javascript
+ * const result = fetchData();
+ * // Log the data if the fetch succeeded.
+ * okDo(console.log);
+ * ```
+ */
+export function okDo<OkData, ErrorMessage>(f: (ok: OkData) => any) {
+  return function(
+    result: Result<OkData, ErrorMessage>
+  ): Result<OkData, ErrorMessage> {
+    if (isOk(result)) {
+      f(result.ok);
+    }
+    return result;
+  };
+}
+
+/**
+ * Runs a function for side effects on the payload, only if the result is Error.
+ *
+ * @param f - the function to run on the error message
+ * @param result - The result to match against
+ * @returns The original result
+ *
+ * ```javascript
+ * const result = fetchData();
+ * // Log an error if the fetch failed.
+ * errorDo(console.error)(result);
+ * ```
+ */
+export function errorDo<OkData, ErrorMessage>(f: (ok: ErrorMessage) => any) {
+  return function(
+    result: Result<OkData, ErrorMessage>
+  ): Result<OkData, ErrorMessage> {
+    if (isError(result)) {
+      f(result.error);
+    }
+    return result;
+  };
+}
+
+/**
+ * Replaces a value that's wrapped in an {ok: data}
+ * Useful if you don't care about the data, just the fact that previous call succeeded.
  *
  * Takes a Result and a mapping function.
  * If the result is an Error, applies the function to the message.
@@ -167,13 +207,40 @@ export function chainError<OkData, ErrorMessage, OkOutput, ErrorOutput>(
  * It wraps the return value in an {error: new_message}.
  * @param f - the function to run on the ok data
  * @param result - The result to match against
+ *
+ * ```javascript
+ * okReplace("fine")(ok(null)) // ok("fine")
+ * ```
  */
-export function replaceError<OkData, ErrorMessage, ErrorOutput>(
-  newError: ErrorOutput
-) {
-  return (
-    result: Result<OkData, ErrorMessage>
-  ): Result<OkData, ErrorOutput> => {
+export function okReplace<OkOutput>(newData: OkOutput) {
+  return <ErrorMessage>(
+    result: Result<any, ErrorMessage>
+  ): Result<OkOutput, ErrorMessage> => {
+    if (isError(result)) {
+      return result;
+    }
+
+    return ok(newData);
+  };
+}
+
+/**
+ * Replaces a message that's wrapped in an `error(message)`
+ * Useful if you don't care about the old message, just the fact that previous call failed.
+ *
+ * Takes a Result and a mapping function.
+ * If the result is an Error, applies the function to the message.
+ * If the result is an Ok, passes the Result through unchanged.
+ *
+ * @param f - the function to run on the ok data
+ * @param result - The result to match against
+ *
+ * ```javascript
+ * errorReplace("something went wrong")(error(null)) // error("something went wrong")
+ * ```
+ */
+export function errorReplace<ErrorOutput>(newError: ErrorOutput) {
+  return <OkData>(result: Result<OkData, any>): Result<OkData, ErrorOutput> => {
     if (isOk(result)) {
       return result;
     }
@@ -183,11 +250,69 @@ export function replaceError<OkData, ErrorMessage, ErrorOutput>(
 }
 
 /**
+ * Get the error message from a result. If it's an Ok, throw an error.
+ * @returns the ok data
+ *
+ * ```typescript
+ * const okResult = ok("good");
+ * okOrThrow(result);
+ * // "good"
+ * ```
+ *
+ * ```typescript
+ * const errorResult = error("bad");
+ * okOrThrow(result);
+ * // throws new Error("bad")
+ * ```
+ */
+export function okOrThrow<OkData>(result: Result<OkData, any>): OkData {
+  if (isError(result)) throw new Error(result.error);
+
+  return result.ok;
+}
+
+/**
+ * Get the error message from a result. If it's an Ok instead, throw an error.
+ * @returns the error message
+ *
+ * ```typescript
+ * errorOrThrow(error("bad"));
+ * // "bad"
+ *
+ * errorOrThrow(ok("good"));
+ * // throws new Error("good")
+ * ```
+ */
+export function errorOrThrow<ErrorMessage>(
+  result: Result<any, ErrorMessage>
+): ErrorMessage {
+  if (isOk(result)) throw new Error(result.ok);
+
+  return result.error;
+}
+
+/**
  * Takes a result, and runs either an onOk or onError function on it.
  * @param onOk - Function to run if the result is an Ok
  * @param onError - Function to run if the result is an Error
  * @param result - Result to match against
  * @returns The return value of the function that gets run.
+ *
+ * ```javascript
+ * // Use `either` to unwrap a Result.
+ * const userId = either(
+ *   await fetchAUser(),
+ *   user => user.id,
+ *   () => null
+ * )
+ *
+ * // Use `either` to act on both cases, but leave wrapped.
+ * const userId = either(
+ *   await fetchAUser(),
+ *   user => ok(user.id),
+ *   () => error("not found")
+ * )
+ * ```
  */
 export function either<OkData, ErrorMessage, OkOutput, ErrorOutput>(
   result: Result<OkData, ErrorMessage>,
@@ -207,51 +332,12 @@ export function either<OkData, ErrorMessage, OkOutput, ErrorOutput>(
  * Converts a result to a boolean.
  * @param result
  * @returns true if Ok, false if Error
+ *
+ * ```javascript
+ * resultToBoolean(ok(1)) // true
+ * resultToBoolean(error(1)) // false
+ * ```
  */
 export function resultToBoolean(result: Result<any, any>): boolean {
   return isOk(result) ? true : false;
-}
-
-/**
- * Get the error message from a result. If it's an Ok, throw an error.
- * @returns the ok data
- *
- * @example
- * ```typescript
- * const okResult = ok("good");
- * okOrThrow(result);
- * // "good"
- *
- * const errorResult = error("bad");
- * okOrThrow(result);
- * // throws new Error("bad")
- * ```
- */
-export function okOrThrow<OkData>(result: Result<OkData, any>): OkData {
-  if (isError(result)) throw new Error(result.error);
-
-  return result.ok;
-}
-
-/**
- * Get the error message from a result. If it's an Ok, throw an error.
- * @returns the error message
- * 
- * @example
- * ```typescript
- * const errorResult = error("bad");
- * errorOrThrow(result);
- * // "bad"
- * 
- * const okResult = ok("good");
- * errorOrThrow(result);
- * // throws new Error("good")
- * ```
- */
-export function errorOrThrow<ErrorMessage>(
-  result: Result<any, ErrorMessage>
-): ErrorMessage {
-  if (isOk(result)) throw new Error(result.ok);
-
-  return result.error;
 }